Changed some Pfd Elements from js classes to pure qml modified:   PfdBackends/airspeedindicatorbackend.cpp modified:   PfdBackends/apinfobackend.h modified:   PfdBackends/bottombarbackend.cpp modified:   PfdBackends/bottombarbackend.h modified:   aircraftManager/definitions/basetypes.h modified:   flightDisplayCompanionPublic.pro modified:   flightDisplayCompanionPublic.pro.user modified:   main.cpp modified:   networkBackend/networkclient.cpp modified:   networkBackend/networkclient.h modified:   networkBackend/networkinterface.cpp modified:   networkBackend/networkinterface.h modified:   networkBackend/readnetworkdata.cpp new file:   qml/images/courseToIcon.png new file:   qml/images/directToIcon.png new file:   qml/images/fdArrow.png modified:   qml/pfdPage/PfdPage.qml deleted:    qml/pfdPage/RadarAltitudeItem.qml modified:   qml/pfdPage/jsHelpers/AirspeedIndicator.js modified:   qml/pfdPage/jsHelpers/Altimeter.js deleted:    qml/pfdPage/jsHelpers/AngleOfAttackIndicator.js deleted:    qml/pfdPage/jsHelpers/AutopilotInfos.js deleted:    qml/pfdPage/jsHelpers/RadioInfos.js deleted:    qml/pfdPage/jsHelpers/WindDataDisplay.js new file:   qml/pfdPage/pfdElements/AngleOfAttackIndicator.qml new file:   qml/pfdPage/pfdElements/AutopilotInfos.qml new file:   qml/pfdPage/pfdElements/BottomBar.qml new file:   qml/pfdPage/pfdElements/ComInfos.qml new file:   qml/pfdPage/pfdElements/NavFreqInfos.qml renamed:    qml/pfdPage/NavInfoItem.qml -> qml/pfdPage/pfdElements/NavInfos.qml new file:   qml/pfdPage/pfdElements/RadarAltitude.qml new file:   qml/pfdPage/pfdElements/WindDataDisplay.qml modified:   qml/qml.qrc modified:   qml/settingsPage/SettingsPage.qml modified:   qml/settingsPage/connection/ConnectionInput.qml modified:   settings/settingsinterface.cpp modified:   settings/settingsinterface.h

+++ networkBackend/networkclient.h
@@ -88,7 +88,7 @@ signals:
     void zulu_secondsChanged(int newValue);
     void local_secondsChanged(int newValue);
     void ground_speedChanged(int newValue);
-    void total_air_tempChanged(int newValue);
+    void total_air_tempChanged(float newValue);
 
     // hsi
     void rotationChanged(double newValue);
